+++ backend/routes/users/authRoutes.js
@@ -2,28 +2,14 @@ const express = require('express');
 const router = express.Router();
 const verifyToken = require('../../utils/verifyToken');
 const {
-  registerUser,
-  loginUser,
   getProtected,
-  refreshToken,
-  login,
   refreshAccessToken
 } = require('../../controller/users/authController');
 
-// POST /api/register
-router.post('/register', registerUser);
-
-// POST /api/login
-router.post('/login', loginUser);
-
-// GET /api/protected
+// GET /auth/protected
 router.get('/protected', verifyToken, getProtected);
 
-// POST /api/refresh
+// POST /auth/refresh
 router.post('/refresh', refreshAccessToken);
 
-// GET
-router.get('/login', login);
-
 module.exports = router;
-
